fixed checkNode bug

+++ horizon/misc_function.py
@@ -32,13 +32,13 @@ def listOfListFLOATtoINT(listOfList):
 
 def checkNodes(nodes, nodes_array=None):
 
-    nodes_vec = np.array(nodes)
+    nodes_vec = np.array(nodes).astype(int)
     # todo check for repeated nodes
     if nodes_array is None:
         checked_nodes = nodes_vec
     else:
         # get from nodes only the nodes active in nodes_array
-        checked_nodes = np.ma.masked_array(nodes_vec, mask=np.logical_not(nodes_array[nodes_vec])).compressed().astype(int)
+        checked_nodes = np.ma.masked_array(nodes_vec, mask=np.logical_not(nodes_array[nodes_vec])).compressed()
 
     # if hasattr(nodes, "__iter__") and not isinstance(nodes, str):
     #     if nodes_array is None:
